Fixed state saving in discover view and added creating of favourite playlist creation

+++ app/src/main/java/com/krykun/movieapp/feature/discover/presentation/DiscoverMoviesState.kt
@@ -2,7 +2,7 @@ package com.krykun.movieapp.feature.discover.presentation
 
 data class DiscoverMoviesState(
     val currentUpcomingPageIndex: Int = -1,
-    val scrollOffsetUpcoming: Float = 0f,
+    val scrollOffsetUpcoming: Int = 0,
     val lastSavedPageUpcoming: Int = 0,
     val loadingState: LoadingState = LoadingState.LOADING
 )
